backline-update: Speed up and provide rational
Avoid `elisp-outline-search', which has no benefit in our case.
We are not interested in changing the appearance of collapsed
top-level expressions (just of real outline headings), so we
don't care if something really is a top-level expression, or
just looks like one.

Avoid `outline-back-to-heading' when `line-beginning-position'
will do.  Also avoid needlessly duplicating work, even though
it would now be cheap enough to not bother.

Also figure out the cause of the `outline-before-first-heading'
signal.  By the time we get to the `re-search-backward', we are
at the very beginning of the buffer and there is nothing before
that.

Since I just spend quite some time trying to re-build this from
first principals: don't.  It's not possible.

In fact after several increasingly desperate attempts I arrived
at a (still broken, but promising) implementation of the same
approach, this package had been using all along.

- `remove-overlays' does not necessarily remove overlays, it may
  also shrink them.  That may work find when implementing some
  new feature, but extending existing code, which relies on this
  function, is ... "interesting".

- Trying to deal with one outline overlay and one backline overlay
  as a unit is futile.  Wiping all backline overlays and then
  adding new companions for all the outline overlays we encounter,
  is the sensible thing to do.

- Trying to figure out what exactly is going on from the arguments
  is also extremely awkward; we get very surprising values.  The
  approach of slightly increasing the search range may same strange,
  but is in fact necessary.  The secret sauce is to give up on `TO'
  as soon as possible and to use `outline-end', to get the real deal.

Hard-code the magical number 1000.  That's not going to change, and
if it does, its easy to revert.  The function `lisp-outline-level'
is the authority on what number we have to use.

One would not expect it, based on the number of commits, but this
is actually the first change since v0.1.0 from seven years ago,
which doesn't just fiddle with metadata and stuff.  Add this text,
so I don't have to figure it all out from scratch again in another
seven years.